requirements.txt
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
opencv-python==4.11.0.86
numpy>=1.23.5,<2.0.0
h5py==3.12.1
keras==2.15.0
tensorboard==2.15.0
torch==2.2.0
torchvision==0.17.0
